mlir: Move ModuleOp into a new `builtin` module
