+++ src/circt/src/lib.rs
@@ -1,9 +1,12 @@
+pub mod builtin;
 pub mod comb;
 pub mod hw;
 pub mod llhd;
 pub mod mlir;
 pub mod seq;
 pub mod std;
+
+pub use builtin::*;
 
 pub mod sys {
     pub use circt_sys::*;
